+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added SPlusUtils with ConvertToInt method taking LowWord/HighWord ushorts
 
 ## [9.2.0] - 2019-03-01
 ### Added